# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Cubit API
  version: "1"
  description: |
    Cubit authenticates against the Israeli Cibus / Pluxee consumer platform,
    reads the meal-card balance, and reports how many fixed-denomination
    vouchers that balance covers.

    Logging in needs a human: Pluxee sends a one-time code out of band, so the
    service parks in `AWAITING_OTP` until someone posts the code back.

        IDLE --login--> AWAITING_OTP --otp--> AUTHENTICATED

    Pluxee also enforces reCAPTCHA on its login endpoint, and its tokens are
    single-use and short-lived, so **cubit cannot log in unattended**. Day to
    day this does not bite: a session obtained once is held encrypted at rest
    and reused across restarts without a new OTP. It matters only when that
    session finally expires, at which point re-authentication is a manual,
    browser-assisted act.

  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: This instance
...
security:
  - apiToken: []
  - basicAuth: []
paths:
  /api/v1/notifications/:
    get:
      tags: [notifications]
      summary: List notification channels
      description: |
        Credentials are redacted in every response: you can see that a token is
        set, never what it is.
      responses:
        "200":
          description: Every configured channel.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Channel" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [notifications]
      summary: Add a notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Channel" }
            examples:
              greenapi:
                summary: GreenAPI (WhatsApp cloud)
                value:
                  name: "my whatsapp"
                  provider: greenapi
                  greenapi:
                    instance_id: "7103"
                    token: "your-greenapi-token"
                    phone: "972501234567"
                  enabled: true
                  notify_on_success: true
                  notify_on_failure: true
              shoutrrr:
                summary: Shoutrrr (Slack, Telegram, Gotify, ntfy, SMTP…)
                value:
                  name: "gotify"
                  provider: shoutrrr
                  shoutrrr: { url: "gotify://hostname/token" }
                  enabled: true
                  notify_on_success: true
              whatsappweb:
                summary: Self-hosted go-whatsapp-web-multidevice
                value:
                  name: "self hosted"
                  provider: whatsapp_web
                  whatsapp_web:
                    base_url: "http://whatsapp:3000"
                    phone: "972501234567"
                  enabled: true
                  notify_on_failure: true
      responses:
        "201":
          description: Created. The response is redacted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }
        "400":
          description: Malformed body, unknown provider, or a missing required field.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/notifications/test:
    post:
      tags: [notifications]
      summary: Send a real test message
      description: |
        Sends using the values in the request **without saving them**, so a
        configuration can be validated before it is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Channel" }
      responses:
        "200":
          description: The provider accepted the message.
        "400":
          description: Malformed body or an incomplete channel.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502":
          description: The provider rejected it or could not be reached.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [notifications]
      summary: Replace a notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Channel" }
      responses:
        "200":
          description: Updated. The response is redacted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }
        "400":
          description: Malformed body or an incomplete channel.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No channel has that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [notifications]
      summary: Remove a notification channel
      responses:
        "204":
          description: Removed.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No channel has that id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/auth/credentials:
    post:
      tags: [auth]
      summary: Supply Cibus credentials at runtime
      description: |
        **This is how you log in.** Post your Cibus username and password here;
        when a `cubit-login` helper is running alongside in watch mode, that is
        the whole trigger. The helper drives a real browser through Pluxee's
        login — the only way past its reCAPTCHA — and Pluxee sends a one-time
        code to the account holder's phone. Submit that code to
        `/api/v1/auth/otp` and the login completes.

        Two calls, then, and nothing else: credentials, then the code.

        With no helper running, this only stores the credentials. Cubit cannot
        log in by itself: Pluxee will not start a login without a captcha token,
        and those are single-use and mintable only by a browser.

        Nothing happens if cubit is already authenticated — there is nothing to
        log in for.

        **Held in memory only.** They are never written to disk, so a restart
        drops them. A restart with a valid persisted session is unaffected: it
        reaches `AUTHENTICATED` with no OTP and no credentials.

        **Refuses to run on an unguarded instance.** Every other route is open
        when no API token is configured, which is deliberate first-run
        behaviour, but an open endpoint that accepts the password to a financial
        account is a different proposition — so this one returns `412` until a
        token is set.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
            examples:
              usual:
                summary: The usual case — username and password only
                value:
                  username: "0501234567"
                  password: "your-cibus-password"
              withCompany:
                summary: Only when the Cibus login screen shows a company field
                value:
                  username: "0501234567"
                  password: "your-cibus-password"
                  company: "acme"
      responses:
        "202":
          description: |
            A login has been started. Pluxee is sending a one-time code to the
            account holder's phone; submit it to `/api/v1/auth/otp`.

            The body never carries the password back, and the password never
            appears in a log line at any level.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "200":
          description: |
            Stored, but nothing is happening: a session is already held, so
            there is nothing to log in for. The credentials will be used by the
            next login.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "400":
          description: |
            Malformed JSON, an unrecognised field, a body over 4 KiB, or either
            half of the pair empty.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A login is awaiting an OTP. Answer it or let it expire first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "412":
          description: No API token is configured, so this endpoint is disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: End the session
      description: |
        Revokes the session at Pluxee — the same `prx_logout` call the web app
        makes when a user signs out — and then clears it locally: cookies out of
        memory, `token.enc` deleted, state back to `IDLE`.

        Revoking upstream is **best effort**. If Pluxee cannot be reached the
        session is still cleared locally and the reason is logged, because a
        backend outage must not leave a live session sitting on disk. The
        upstream call is attempted first, since after the local wipe there are
        no cookies left to revoke with.

        Idempotent: logging out with nothing held is a success, so a caller can
        always reach a known state without checking first. Nothing is sent
        upstream in that case — there is nothing to revoke.
      responses:
        "204":
          description: |
            Logged out. Returned even when the upstream revocation failed; the
            local session is gone either way.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502":
          description: The logout could not be completed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/otp:
    post:
      tags: [auth]
      summary: Submit the OTP code
      description: |
        The second and final step of a login. Post the code the account holder
        received by SMS.

        A `202` here is the normal, successful answer for a browser-assisted
        login: the code has been received and handed to the helper, which
        completes the login in the browser it is holding open. Watch
        `/api/v1/auth/status` for the state to reach `AUTHENTICATED`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OTPRequest" }
      responses:
        "200":
          description: Authenticated. The session is persisted, encrypted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "202":
          description: |
            The code was received and parked for a browser-assisted login to
            complete. Returned when the login was started by the `cubit-login`
            helper rather than by cubit: the helper holds the challenge and can
            mint the captcha token the submission needs, so cubit relays the
            code instead of submitting it.

            This is the response a phone-side auto-forwarder should expect.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "400":
          description: Malformed body, an unrecognised field, or an empty code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: |
            Wrong code. `attempts_remaining` says how many are left before the
            challenge is torn down and a fresh login is required.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "409":
          description: |
            No OTP is pending. Either no login has been started, or one already
            succeeded — the `message` says which, and `state` tells you plainly.
            Posting a code twice lands here, and is harmless.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "410":
          description: The challenge outlived its TTL. Start a new login.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "502": { $ref: "#/components/responses/BadGateway" }
  /api/v1/auth/status:
    get:
      tags: [auth]
      summary: Current authentication state
      responses:
        "200":
          description: The state machine's current position.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/balance:
    get:
      tags: [balance]
      summary: Balance and voucher calculation
      description: |
        All money is handled as `int64` agorot; the decimal strings alongside
        each figure are for display only.
      responses:
        "200":
          description: The current balance and what it buys.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503":
          description: |
            Not authenticated. The body carries the auth state, so the caller
            can tell whether an OTP is pending or credentials are missing.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
components:
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: |
        Guards `/api/v1` and `/metrics`. `/healthz`, `/readyz` and `/api/docs`
        are always open.

        **Bootstrap mode:** when no token is configured, `/api/v1` and
        `/metrics` are open to anyone who can reach the port. That is deliberate
        first-run behaviour — the service must be reachable before it is
        configured — and the process logs a warning at startup while it lasts.
        `POST /api/v1/auth/credentials` is the one exception: it refuses to
        operate until a token is set.
    basicAuth:
      type: http
      scheme: basic
      description: |
        The same token, presented as the password half of basic auth. The
        username is ignored. Comparison is constant-time.
  responses:
    Unauthorized:
      description: An API token is configured and the request did not carry it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: The Pluxee backend could not be reached.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthStatus" }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
    CredentialsRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Cibus / Pluxee username. Surrounding whitespace is trimmed.
          example: alice
        password:
          type: string
          format: password
          description: Cibus / Pluxee password.
        company:
          type: string
          description: |
            **Optional — omit it entirely unless the Cibus login screen actually
            shows a company field.** Most accounts do not need one; verified
            against a real account on 2026-09-09 that logging in with no company
            succeeds.

            Do not try to determine this by calling `/auth/showCompanyField`
            yourself without a reCAPTCHA token: it answers `true` without one and
            `false` with one, for the same username.
      additionalProperties: false
    OTPRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: The one-time code delivered by SMS or voice call.
          example: "123456"
      additionalProperties: false
...